[!*] [unquote_string] Fixed bug in deserealization of lua string.
Found and fixed very interesting bug in deserealization of
linear lua string like

  font = "Fonts\\2002.ttf"

Initially the code changed "\202" to ASCII(202) and replace
"\\" -> "\" not occured.

So we have to exclude "\\" sequences before other deserialization.
Problem is that there no spare characters to substityte "\\" to it.
(As string may contain any possible character.) So string is splitted
to parts by delimiter "\\", that parts deserialized in normal way and
later concatted with "\" delimiter.