[*] [lua_code_formatter] Fixed performance bug.
Problem was that formatter was (incorrectly) initialized as

  printer = request('^.^.^.string.text_block.interface')

Later in variate we make trial copy

  local printer_class = request('^.^.^.string.text_block.interface')
  local trial_presentation = new(printer_class)

in hope that <printer_class> is empty stub. But indeed this
contatins constanly growing pile of lines, so cloning it with new()
requires O(N) time.

Fixed this by correcly initializing field as

  printer = new(request('^.^.^.string.text_block.interface'))

Morale: always clone class before use if it may have copies.