[!/] [formats.lua_table_code] Reimplemented lua table serializer.
[-] Dropped previous versions of [lua_table] and [lua_table_code]
and substituted new.

New functionality will going to be SLOWER. But anyway I'll choose it
because it's SIMPLER. Simpler to write and understand.

It was not a quick task to implement this change as [lua_table_code]
used internal parts of [lua_table] with parallel output serialization.
Now arbitrary table serialization is divided in two consecutive
stages:

  1. Construct tree of tokens to serialize.
  2. Serialize tree using standalone token serializers.

--

Well, it's a LOT slower. Almost two times.

file_size |  old  |  new
----------+-------+-------
 28.1MiB  | 12.1s | 22.1s